--- server/routes/customDomainsRoutes.ts
@@ -1,7 +1,0 @@
-import { Router } from 'express';
-
-const router = Router();
-
-// TODO: Implement custom domain routes
-
-export default router;
