+++ server/routes/index.ts
@@ -51,19 +51,17 @@ router.use('/analytics', analyticsRoutes);
 
 // Mount admin and feature routes
 router.use('/admin', adminRoutes);
+import customDomainsRoutes from './customDomains.js';
 import collaborationRoutes from './collaborationRoutes.js';
-import customDomainsRoutes from './customDomainsRoutes.js';
-import whiteLabelRoutes from './whiteLabelRoutes.js';
 
 router.use('/calendar', calendarRoutes);
-router.use('/collaboration', collaborationRoutes); // Now defined
 router.use('/approvals', approvalRoutes);
 router.use('/expenses', expenseRoutes);
 router.use('/reporting', reportingRoutes);
 router.use('/organization-funding', organizationFundingRoutes);
 router.use('/stripe', stripeOAuthRoutes);
-router.use('/custom-domains', customDomainsRoutes); // Now defined
-router.use('/white-label', whiteLabelRoutes); // Now defined
+router.use('/custom-domains', customDomainsRoutes);
+router.use('/collaboration', collaborationRoutes);
 
 
 // Mount test routes in development only
@@ -82,7 +80,7 @@ router.use('/flights', flightRoutes);
 router.use('/export', exportRoutes);
 
 // Import and register simplified white label routes
-import { registerSimplifiedWhiteLabelRoutes } from './whiteLabelSimplified.js';
+import { registerWhiteLabelRoutes } from './whiteLabelRoutes.js';
 // Note: This will be handled in the main server file since it needs the app instance
 // router.use('/todos', todosRoutes);
 // router.use('/notes', notesRoutes);
@@ -388,7 +386,7 @@ router.get('/health', (req, res) => {
 export function registerDirectRoutes(app: any) {
   registerAdminSettingsRoutes(app);
   registerWhiteLabelStatusRoutes(app);
-  registerSimplifiedWhiteLabelRoutes(app); // Call the imported function
+  registerWhiteLabelRoutes(app); // Call the imported function
   registerBookingRoutes(app);
   registerCorporateCardRoutes(app);
 }
